split the timeline model and fold its duplicated handlers together
feed and thread mode each carried their own copy of the like, preview, and
post-action handling, so a fix in one never reached the other. they now
share one set of methods, and thread and reply mode move to their own
files: model.go 1327 -> 908 lines.

the like handler in thread mode now also keeps the selection on screen,
matching the feed.
